+++ src/main/java/com/ideas2it/employee/service/impl/ProjectManagementServiceImpl.java
@@ -16,7 +16,6 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 
 /**
@@ -25,26 +24,34 @@ import org.springframework.stereotype.Component;
  * @version  4.1 10-10-2022.
  * @author  Ananth K.
  */
-@Component("projectService")
+@Component
 public class ProjectManagementServiceImpl implements ProjectManagementService {
-    
+
 	@Autowired
     private ProjectDao projectDao;
-	@Autowired
-	private ApplicationContext context;
 	
+	private EmployeeManagementService service;
+
+	public EmployeeManagementService getService() {
+		return service;
+	}
+    
+	public void setService(EmployeeManagementServiceImpl service) {
+		this.service = service;
+	}
+
 	private Logger logger = LogManager.getLogger(ProjectManagementServiceImpl.class);
     
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public ProjectDTO addProject(ProjectDTO projectDto) {
-        Project project = ProjectMapper.toProject(projectDto);
-        ProjectDTO projectDTO = ProjectMapper.toProjectDto(projectDao.save(project));
-        logger.info("Project created succuessfully ProjectID =" + projectDto.getId());
-        return projectDTO;
-    }
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public ProjectDTO addProject(ProjectDTO projectDto) {
+		Project project = ProjectMapper.toProject(projectDto);
+		ProjectDTO projectDTO = ProjectMapper.toProjectDto(projectDao.save(project));
+		logger.info("Project created succuessfully ProjectID =" + projectDto.getId());
+		return projectDTO;
+	}
 
 	/**
 	 * {@inheritDoc}
@@ -78,14 +85,8 @@ public class ProjectManagementServiceImpl implements ProjectManagementService {
 		List<Project> projects = projectDao.findByName(name);
 		List<ProjectDTO> projectDtos = new ArrayList<ProjectDTO>();
 
-		if (!projects.isEmpty()) {
-			for (Project project : projects) {
-
-				projectDtos.add(ProjectMapper.toProjectDto(project));
-			}
-		} else {
-			logger.error(Constant.PROJECT_NOT_FOUND);
-			throw new EMSException(Constant.PROJECT_NOT_FOUND, Constant.ERROR_CODE104);
+		for (Project project : projects) {
+			projectDtos.add(ProjectMapper.toProjectDto(project));
 		}
 		return projectDtos;
 	}
@@ -95,17 +96,21 @@ public class ProjectManagementServiceImpl implements ProjectManagementService {
 	 */
 	@Override
 	public void deleteProject(int projectId) {
-		projectDao.deleteById(projectId);
+		
+		if (null != projectExists(projectId)) {
+			projectDao.deleteById(projectId);
+		} else {
+			logger.error(Constant.PROJECT_NOT_FOUND + projectId);
+			throw new EMSException(Constant.PROJECT_NOT_FOUND, Constant.ERROR_CODE107);
+		}
 		logger.info("Project deleted successfully ProjectID =" + projectId);
 	}
 
 	@Override
 	public ProjectDTO assignEmployee(int employeeId, int projectId) {
 
-		EmployeeManagementService service = (EmployeeManagementServiceImpl) 
-				context.getBean("employeeService");
-		Employee employee = ProjectMapper.toEmployee(service.employeeExists(employeeId));
-		Project project = ProjectMapper.toProject(projectExists(projectId));
+		Employee employee = service.employeeExists(employeeId);
+		Project project = projectExists(projectId);
 		ProjectDTO projectDto = null;
 
 		try {
@@ -123,14 +128,12 @@ public class ProjectManagementServiceImpl implements ProjectManagementService {
 		return projectDto;
 	}
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public ProjectDTO projectExists(int projectId) {
-        List<ProjectDTO> projectDtos = getAllProjects();
-        ProjectDTO projectDto = projectDtos.stream().filter(x -> x.getId() == (projectId))
-                          .findFirst().orElse(null);
-        return projectDto;
-    }
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public Project projectExists(int projectId) {
+
+		return projectDao.findById(projectId).orElse(null);
+	}
 }
